Fix: Remove namespace configuration to resolve launch file error
The current turtlebot4_bringup/launch/lite.launch.py file was throwing an error as the namespace argument was not passed. This has been resolved by removing the namespace configuration in the generate_launch_description function. The namespace is already specified in the parameters file, making the previous configuration redundant. This change not only fixes the error but also simplifies the code and avoids potential conflicts between the function and parameters file."